refactoring the diagnostic settings helper
