+++ src/steps/resource-manager/utils/__recordings__/create-diagnostic-settings-entities-and-relationships_4070255659/recording.har
